llama_aws_eks_vllm_grpc_ttft_radha/docker/requirements.txt
vllm
torch
transformers
protobuf
tokenizers
prometheus_client
grpcio==1.65.5
grpcio-tools==1.65.5
grpcio-reflection==1.65.5
grpcio-health-checking==1.65.5
